+++ src/main/java/org/corpus_tools/annis/gui/graphml/DocumentGraphMapper.java
@@ -15,7 +15,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.SortedMap;
 import java.util.TreeMap;
+import java.util.stream.Collectors;
 import javax.xml.namespace.QName;
 import javax.xml.stream.XMLEventReader;
 import javax.xml.stream.XMLStreamException;
@@ -56,23 +58,21 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
   private final class RecreateTextForRootNodeTraverser implements GraphTraverseHandler {
     private final Map<SToken, Range<Integer>> token2Range;
     private final StringBuilder text;
-    private final String textName;
 
     private RecreateTextForRootNodeTraverser(Map<SToken, Range<Integer>> token2Range,
-        StringBuilder text, String textName) {
+        StringBuilder text) {
       this.token2Range = token2Range;
       this.text = text;
-      this.textName = textName;
     }
 
     @SuppressWarnings("rawtypes")
     @Override
-    public boolean checkConstraint(SGraph.GRAPH_TRAVERSE_TYPE traversalType,
-        String traversalId, SRelation relation, SNode currNode, long order) {
+    public boolean checkConstraint(SGraph.GRAPH_TRAVERSE_TYPE traversalType, String traversalId,
+        SRelation relation, SNode currNode, long order) {
       if (relation == null) {
         return true;
       } else if (relation instanceof SOrderRelation
-          && Objects.equal(textName, relation.getType())) {
+          && Objects.equal("", relation.getType())) {
         return true;
       } else {
         return false;
@@ -88,8 +88,7 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
         SNode currNode, SRelation<SNode, SNode> relation, SNode fromNode, long order) {
 
       if (currNode instanceof SToken) {
-        SFeature featTokWhitespaceBefore =
-            currNode.getFeature("annis::tok-whitespace-before");
+        SFeature featTokWhitespaceBefore = currNode.getFeature("annis::tok-whitespace-before");
         if (featTokWhitespaceBefore != null) {
           text.append(featTokWhitespaceBefore.getValue().toString());
         }
@@ -177,7 +176,8 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
     Optional<String> currentTargetId = Optional.empty();
     Optional<String> currentComponent = Optional.empty();
 
-    Map<String, STextualDS> datasources = new HashMap<>();
+    SortedMap<String, STextualDS> datasources = new TreeMap<>();
+    Map<SToken, SToken> gapEdges = new HashMap<>();
 
     Map<String, String> data = new HashMap<>();
 
@@ -205,7 +205,6 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
                 if (id != null) {
                   currentNodeId = Optional.ofNullable(id.getValue());
                 }
-
               }
               break;
             case "edge":
@@ -269,7 +268,7 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
               if (currentSourceId.isPresent() && currentTargetId.isPresent()
                   && currentComponent.isPresent()) {
                 mapAndAddEdge(currentSourceId.get(), currentTargetId.get(), currentComponent.get(),
-                    data);
+                    data, gapEdges);
               }
 
               currentSourceId = Optional.empty();
@@ -295,8 +294,29 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
     }
 
 
-    // find all chains of SOrderRelations and reconstruct the texts belonging to
-    // them
+    // Create the textual DS for the minimal token roots
+    if (datasources.isEmpty()) {
+      // Find roots ignoring the data source
+      List<SToken> orderedTokenRoots = getOrderedTokenRootsForDatasource(graph, gapEdges, null);
+      if (!orderedTokenRoots.isEmpty()) {
+        // Create an empty default data source, because there are none yet
+        STextualDS ds = SaltFactory.createSTextualDS();
+        ds.setText("");
+        graph.addNode(ds);
+        // Use this default data source to attach the token to
+        recreateTextForRootNodes(orderedTokenRoots, ds);
+      }
+    } else {
+      // Sort data sources by their name
+
+      for (STextualDS ds : datasources.values()) {
+        List<SToken> orderedTokenRoots = getOrderedTokenRootsForDatasource(graph, gapEdges, ds);
+        recreateTextForRootNodes(orderedTokenRoots, ds);
+      }
+    }
+
+
+    // Create the text annotation for the segmentation nodes
     Multimap<String, SNode> orderRoots = graph.getRootsByRelationType(SALT_TYPE.SORDER_RELATION);
     if (orderRoots.isEmpty() && graph.getTokens().size() == 1) {
       // if there is only one token, there won't be any order relations
@@ -307,25 +327,7 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
       if (SaltUtil.SALT_NULL_VALUE.equals(name)) {
         name = null;
       }
-      if (name == null || "".equals(name)) {
-        // Decide to which data sources the token belong to. A single data source could have
-        // multiple root token due to gaps in the result. But there also could be multiple root
-        // token because multiple textual data sources are involved.
-        Multimap<STextualDS, SNode> rootsForDatasource = HashMultimap.create();
-        for (SNode r : roots) {
-          for (String targetId : isPartOf.get(r.getId())) {
-            STextualDS targetDs = datasources.get(targetId);
-            if (targetDs != null) {
-              rootsForDatasource.put(targetDs, r);
-            }
-          }
-        }
-
-        for (STextualDS ds : rootsForDatasource.keySet()) {
-          // re-create text if this is the default (possible virtual) tokenization
-          recreateTextForRootNodes(name, rootsForDatasource.get(ds), ds);
-        }
-      } else {
+      if (name != null && !"".equals(name)) {
         // add the text as label to the spans
         addTextToSegmentation(name, roots);
       }
@@ -333,7 +335,6 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
 
     addNodeLayers();
   }
-
 
 
   /**
@@ -370,7 +371,7 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
   }
 
   private void mapAndAddEdge(String sourceId, String targetId, String componentRaw,
-      Map<String, String> labels) {
+      Map<String, String> labels, Map<SToken, SToken> gapEdges) {
 
     SNode source = graph.getNode(Helper.addSaltPrefix(sourceId));
     SNode target = graph.getNode(Helper.addSaltPrefix(targetId));
@@ -397,7 +398,15 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
           rel = graph.createRelation(source, target, SALT_TYPE.SPOINTING_RELATION, null);
           break;
         case ORDERING:
-          rel = graph.createRelation(source, target, SALT_TYPE.SORDER_RELATION, null);
+          if ("annis".equals(component.getLayer())
+              && "datasource-gap".equals(component.getName())) {
+            if (source instanceof SToken && target instanceof SToken) {
+              gapEdges.put((SToken) source, (SToken) target);
+            }
+          } else {
+            rel = graph.createRelation(source, target, SALT_TYPE.SORDER_RELATION, null);
+          }
+
           break;
         case COVERAGE:
           // only add coverage edges in salt to spans, not structures
@@ -419,25 +428,64 @@ public class DocumentGraphMapper extends AbstractGraphMLMapper {
     }
   }
 
-  private void recreateTextForRootNodes(final String textName,
-      Collection<? extends SNode> rootsForText, STextualDS originalDs) {
+  private List<SToken> getOrderedTokenRootsForDatasource(SDocumentGraph graph,
+      Map<SToken, SToken> gapEdges, STextualDS ds) {
+    Map<SToken, SToken> outgoingOrderingEdges = new HashMap<>();
+    Map<SToken, SToken> incomingOrderingEdgesWithGaps = new HashMap<>();
+
+    for (SOrderRelation rel : graph.getOrderRelations()) {
+      if (("".equals(rel.getType()) || rel.getType() == null) && rel.getSource() instanceof SToken
+          && rel.getTarget() instanceof SToken) {
+        outgoingOrderingEdges.put((SToken) rel.getSource(), (SToken) rel.getTarget());
+        incomingOrderingEdgesWithGaps.put((SToken) rel.getTarget(), (SToken) rel.getSource());
+      }
+    }
+
+    for (Map.Entry<SToken, SToken> rel : gapEdges.entrySet()) {
+      incomingOrderingEdgesWithGaps.put(rel.getValue(), rel.getKey());
+    }
+    
+    // Get all root nodes of this data source (token without any incoming ordering edge)
+    List<SToken> datasourceRoots = graph.getTokens().stream()
+        .filter(t -> !incomingOrderingEdgesWithGaps.containsKey(t))
+        .filter(t -> ds == null
+            || isPartOf.get(t.getId()).stream().anyMatch(dsId -> ds.getId().equals(dsId)))
+        .collect(Collectors.toList());
+
+    List<SToken> result = new ArrayList<>();
+    // Create an ordered list of local roots (without the gap ordering edges) by following the
+    // outgoing ordering edges
+    for(SToken root : datasourceRoots) {
+      SToken token = root;
+
+      while (token != null) {
+        result.add(token);
+
+        // Skip to to end of this token chain
+        while (outgoingOrderingEdges.containsKey(token)) {
+          token = outgoingOrderingEdges.get(token);
+        }
+        // Step over the possible gap
+        token = gapEdges.get(token);
+      }
+    }
+
+    return result;
+
+  }
+
+
+  private void recreateTextForRootNodes(Collection<? extends SNode> rootsForText, STextualDS ds) {
 
     final StringBuilder text = new StringBuilder();
-    final STextualDS ds;
-    if (originalDs == null) {
-      ds = graph.createTextualDS("");
-      ds.setName(textName);
-    } else {
-      ds = originalDs;
-    }
 
     Map<SToken, Range<Integer>> token2Range = new HashMap<>();
 
     // traverse the token chain using the order relations
     for (SNode root : rootsForText) {
       graph.traverse(Arrays.asList(root), SGraph.GRAPH_TRAVERSE_TYPE.TOP_DOWN_DEPTH_FIRST,
-          "ORDERING_" + textName,
-          new RecreateTextForRootNodeTraverser(token2Range, text, textName));
+          "ORDERING",
+          new RecreateTextForRootNodeTraverser(token2Range, text));
     }
 
     // update the actual text